:memo:Alteração

+++ produto-ms/src/main/java/br/com/tech4me/produtoms/view/model/ProdutoModeloAlteracao.java
@@ -4,7 +4,7 @@ public class ProdutoModeloAlteracao {
     private String id;
     private String nome;
     private String codigo;
-    private float preco;
+    private double preco;
     private Integer estoque;
     
     public String getId() {
@@ -25,10 +25,10 @@ public class ProdutoModeloAlteracao {
     public void setCodigo(String codigo) {
         this.codigo = codigo;
     }
-    public float getPreco() {
+    public double getPreco() {
         return preco;
     }
-    public void setPreco(float preco) {
+    public void setPreco(double preco) {
         this.preco = preco;
     }
     public Integer getEstoque() {
